+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComponents.java
@@ -116,8 +116,8 @@ public final class EntityComponents {
 		if (attribute instanceof ForeignKey) {
 			return (ComponentValueBuilder<C, T, B>) textField((ForeignKey) attribute);
 		}
-		ValueAttributeDefinition<T> attributeDefinition = definition(attribute);
-		if (!attributeDefinition.items().isEmpty()) {
+		AttributeDefinition<T> attributeDefinition = definition(attribute);
+		if (itemBased(attributeDefinition)) {
 			return (ComponentValueBuilder<C, T, B>) itemComboBox(attribute);
 		}
 		Attribute.Type<T> type = attribute.type();
@@ -128,7 +128,7 @@ public final class EntityComponents {
 			return (ComponentValueBuilder<C, T, B>) textField(attribute);
 		}
 		if (type.isBoolean()) {
-			if (attributeDefinition.nullable()) {
+			if (nullable(attributeDefinition)) {
 				return (ComponentValueBuilder<C, T, B>) nullableCheckBox((Attribute<Boolean>) attribute);
 			}
 
@@ -150,7 +150,7 @@ public final class EntityComponents {
 			return (ComponentValueBuilder<C, T, B>) bigDecimalField((Attribute<BigDecimal>) attribute);
 		}
 		if (type.isEnum()) {
-			return (ComponentValueBuilder<C, T, B>) comboBox(attribute, createEnumComboBoxModel(attribute, attributeDefinition.nullable()));
+			return (ComponentValueBuilder<C, T, B>) comboBox(attribute, createEnumComboBoxModel(attribute, nullable(attributeDefinition)));
 		}
 		if (attribute.type().isByteArray()) {
 			return (ComponentValueBuilder<C, T, B>) byteArrayInputPanel((Attribute<byte[]>) attribute);
@@ -302,15 +302,16 @@ public final class EntityComponents {
 	 * @throws IllegalArgumentException in case the given attribute has no associated items
 	 */
 	public <T> ItemComboBoxBuilder<T> itemComboBox(Attribute<T> attribute) {
-		ValueAttributeDefinition<T> attributeDefinition = definition(attribute);
-		if (attributeDefinition.items().isEmpty()) {
+		AttributeDefinition<T> attributeDefinition = definition(attribute);
+		if (!itemBased(attributeDefinition)) {
 			throw new IllegalArgumentException("Attribute '" + attributeDefinition.attribute() + "' is not a item based attribute");
 		}
+		ValueAttributeDefinition<T> valueAttributeDefinition = (ValueAttributeDefinition<T>) attributeDefinition;
 
 		return Components.itemComboBox()
-						.items(attributeDefinition.items())
-						.toolTipText(attributeDefinition.description().orElse(null))
-						.nullable(attributeDefinition.nullable());
+						.items(valueAttributeDefinition.items())
+						.toolTipText(valueAttributeDefinition.description().orElse(null))
+						.nullable(valueAttributeDefinition.nullable());
 	}
 
 	/**
@@ -324,7 +325,7 @@ public final class EntityComponents {
 	 */
 	public <T, C extends JComboBox<T>, B extends ComboBoxBuilder<C, T, B>> ComboBoxBuilder<C, T, B> comboBox(Attribute<T> attribute,
 																																																					 ComboBoxModel<T> comboBoxModel) {
-		ValueAttributeDefinition<T> attributeDefinition = definition(attribute);
+		AttributeDefinition<T> attributeDefinition = definition(attribute);
 
 		return (ComboBoxBuilder<C, T, B>) Components.comboBox()
 						.model(comboBoxModel)
@@ -338,7 +339,7 @@ public final class EntityComponents {
 	 * @return a {@link TemporalFieldPanel} builder
 	 */
 	public <T extends Temporal> TemporalFieldPanel.Builder<T> temporalFieldPanel(Attribute<T> attribute) {
-		ValueAttributeDefinition<T> attributeDefinition = definition(attribute);
+		AttributeDefinition<T> attributeDefinition = definition(attribute);
 
 		return Components.temporalFieldPanel()
 						.temporalClass(attribute.type().valueClass())
@@ -353,11 +354,11 @@ public final class EntityComponents {
 	 * @return a {@link TextFieldPanel} builder
 	 */
 	public TextFieldPanel.Builder textFieldPanel(Attribute<String> attribute) {
-		ValueAttributeDefinition<String> attributeDefinition = definition(attribute);
+		AttributeDefinition<String> attributeDefinition = definition(attribute);
 
 		return Components.textFieldPanel()
 						.toolTipText(attributeDefinition.description().orElse(null))
-						.maximumLength(attributeDefinition.maximumLength())
+						.maximumLength(maximumLength(attributeDefinition))
 						.dialogTitle(attributeDefinition.caption())
 						.buttonIcon(ICONS.editText().large());
 	}
@@ -368,14 +369,14 @@ public final class EntityComponents {
 	 * @return a JTextArea builder
 	 */
 	public TextAreaBuilder textArea(Attribute<String> attribute) {
-		ValueAttributeDefinition<String> attributeDefinition = definition(attribute);
+		AttributeDefinition<String> attributeDefinition = definition(attribute);
 		if (!attribute.type().isString()) {
 			throw new IllegalArgumentException("Cannot create a text area for a non-string attribute");
 		}
 
 		return Components.textArea()
 						.toolTipText(attributeDefinition.description().orElse(null))
-						.maximumLength(attributeDefinition.maximumLength());
+						.maximumLength(maximumLength(attributeDefinition));
 	}
 
 	/**
@@ -387,9 +388,9 @@ public final class EntityComponents {
 	 * @return a JTextField builder
 	 */
 	public <T, C extends JTextField, B extends TextFieldBuilder<C, T, B>> TextFieldBuilder<C, T, B> textField(Attribute<T> attribute) {
-		ValueAttributeDefinition<T> attributeDefinition = definition(attribute);
+		AttributeDefinition<T> attributeDefinition = definition(attribute);
 
-		if (!attributeDefinition.items().isEmpty()) {
+		if (itemBased(attributeDefinition)) {
 			return (TextFieldBuilder<C, T, B>) Components.textField()
 							.valueClass(attribute.type().valueClass())
 							.format(new ItemReadOnlyFormat(attributeDefinition))
@@ -413,7 +414,7 @@ public final class EntityComponents {
 		return (TextFieldBuilder<C, T, B>) Components.textField()
 						.valueClass(attribute.type().valueClass())
 						.format(attributeDefinition.format().orElse(null))
-						.maximumLength(attributeDefinition.maximumLength())
+						.maximumLength(maximumLength(attributeDefinition))
 						.toolTipText(attributeDefinition.description().orElse(null));
 	}
 
@@ -424,7 +425,7 @@ public final class EntityComponents {
 	 * @return a {@link TemporalField} builder
 	 */
 	public <T extends Temporal> TemporalField.Builder<T> temporalField(Attribute<T> attribute) {
-		ValueAttributeDefinition<T> attributeDefinition = definition(attribute);
+		AttributeDefinition<T> attributeDefinition = definition(attribute);
 
 		return Components.temporalField()
 						.temporalClass(attributeDefinition.attribute().type().valueClass())
@@ -439,7 +440,7 @@ public final class EntityComponents {
 	 * @return a {@link NumberField} builder
 	 */
 	public NumberField.Builder<Short> shortField(Attribute<Short> attribute) {
-		ValueAttributeDefinition<Short> attributeDefinition = definition(attribute);
+		AttributeDefinition<Short> attributeDefinition = definition(attribute);
 
 		return Components.shortField()
 						.format(attributeDefinition.format().orElse(null))
@@ -452,7 +453,7 @@ public final class EntityComponents {
 	 * @return a {@link NumberField} builder
 	 */
 	public NumberField.Builder<Integer> integerField(Attribute<Integer> attribute) {
-		ValueAttributeDefinition<Integer> attributeDefinition = definition(attribute);
+		AttributeDefinition<Integer> attributeDefinition = definition(attribute);
 
 		return Components.integerField()
 						.format(attributeDefinition.format().orElse(null))
@@ -465,7 +466,7 @@ public final class EntityComponents {
 	 * @return a {@link NumberField} builder
 	 */
 	public NumberField.Builder<Long> longField(Attribute<Long> attribute) {
-		ValueAttributeDefinition<Long> attributeDefinition = definition(attribute);
+		AttributeDefinition<Long> attributeDefinition = definition(attribute);
 
 		return Components.longField()
 						.format(attributeDefinition.format().orElse(null))
@@ -478,7 +479,7 @@ public final class EntityComponents {
 	 * @return a {@link NumberField} builder
 	 */
 	public NumberField.Builder<Double> doubleField(Attribute<Double> attribute) {
-		ValueAttributeDefinition<Double> attributeDefinition = definition(attribute);
+		AttributeDefinition<Double> attributeDefinition = definition(attribute);
 
 		return Components.doubleField()
 						.format(attributeDefinition.format().orElse(null))
@@ -492,7 +493,7 @@ public final class EntityComponents {
 	 * @return a {@link NumberField} builder
 	 */
 	public NumberField.Builder<BigDecimal> bigDecimalField(Attribute<BigDecimal> attribute) {
-		ValueAttributeDefinition<BigDecimal> attributeDefinition = definition(attribute);
+		AttributeDefinition<BigDecimal> attributeDefinition = definition(attribute);
 
 		return Components.bigDecimalField()
 						.format(attributeDefinition.format().orElse(null))
@@ -507,9 +508,9 @@ public final class EntityComponents {
 	 * @return a {@link javax.swing.JSlider} builder
 	 */
 	public SliderBuilder slider(Attribute<Integer> attribute) {
-		ValueAttributeDefinition<Integer> attributeDefinition = definition(attribute);
-		Number minimumValue = attributeDefinition.minimum().orElse(null);
-		Number maximumValue = attributeDefinition.maximum().orElse(null);
+		AttributeDefinition<Integer> attributeDefinition = definition(attribute);
+		Number minimumValue = minimumValue(attributeDefinition);
+		Number maximumValue = maximumValue(attributeDefinition);
 		if (minimumValue == null || maximumValue == null) {
 			throw new IllegalArgumentException("Cannot create a slider for an attribute without min and max values");
 		}
@@ -528,7 +529,7 @@ public final class EntityComponents {
 	 * @return a {@link javax.swing.JSlider} builder
 	 */
 	public SliderBuilder slider(Attribute<Integer> attribute, BoundedRangeModel boundedRangeModel) {
-		ValueAttributeDefinition<Integer> attributeDefinition = definition(attribute);
+		AttributeDefinition<Integer> attributeDefinition = definition(attribute);
 
 		return Components.slider()
 						.model(boundedRangeModel)
@@ -541,9 +542,9 @@ public final class EntityComponents {
 	 * @return a {@link javax.swing.JSpinner} builder
 	 */
 	public NumberSpinnerBuilder<Integer> integerSpinner(Attribute<Integer> attribute) {
-		ValueAttributeDefinition<Integer> attributeDefinition = definition(attribute);
-		Number minimumValue = attributeDefinition.minimum().orElse(null);
-		Number maximumValue = attributeDefinition.maximum().orElse(null);
+		AttributeDefinition<Integer> attributeDefinition = definition(attribute);
+		Number minimumValue = minimumValue(attributeDefinition);
+		Number maximumValue = maximumValue(attributeDefinition);
 
 		return Components.integerSpinner()
 						.minimum(minimumValue == null ? null : minimumValue.intValue())
@@ -557,9 +558,9 @@ public final class EntityComponents {
 	 * @return a {@link javax.swing.JSpinner} builder
 	 */
 	public NumberSpinnerBuilder<Double> doubleSpinner(Attribute<Double> attribute) {
-		ValueAttributeDefinition<Double> attributeDefinition = definition(attribute);
-		Number minimumValue = attributeDefinition.minimum().orElse(null);
-		Number maximumValue = attributeDefinition.maximum().orElse(null);
+		AttributeDefinition<Double> attributeDefinition = definition(attribute);
+		Number minimumValue = minimumValue(attributeDefinition);
+		Number maximumValue = maximumValue(attributeDefinition);
 
 		return Components.doubleSpinner()
 						.minimum(minimumValue == null ? null : minimumValue.doubleValue())
@@ -575,7 +576,7 @@ public final class EntityComponents {
 	 * @return a {@link javax.swing.JSpinner} builder
 	 */
 	public <T> ListSpinnerBuilder<T> listSpinner(Attribute<T> attribute, SpinnerListModel listModel) {
-		ValueAttributeDefinition<T> attributeDefinition = definition(attribute);
+		AttributeDefinition<T> attributeDefinition = definition(attribute);
 
 		return Components.<T>listSpinner()
 						.model(listModel)
@@ -589,14 +590,15 @@ public final class EntityComponents {
 	 * @return a {@link javax.swing.JSpinner} builder
 	 */
 	public <T> ItemSpinnerBuilder<T> itemSpinner(Attribute<T> attribute) {
-		ValueAttributeDefinition<T> attributeDefinition = definition(attribute);
-		if (attributeDefinition.items().isEmpty()) {
+		AttributeDefinition<T> attributeDefinition = definition(attribute);
+		if (!itemBased(attributeDefinition)) {
 			throw new IllegalArgumentException("Attribute '" + attributeDefinition.attribute() + "' is not a item based attribute");
 		}
+		ValueAttributeDefinition<T> valueAttributeDefinition = (ValueAttributeDefinition<T>) attributeDefinition;
 
 		return Components.<T>itemSpinner()
-						.model(new SpinnerListModel(attributeDefinition.items()))
-						.toolTipText(attributeDefinition.description().orElse(null));
+						.model(new SpinnerListModel(valueAttributeDefinition.items()))
+						.toolTipText(valueAttributeDefinition.description().orElse(null));
 	}
 
 	/**
@@ -605,7 +607,7 @@ public final class EntityComponents {
 	 * @return a JFormattedTextField builder
 	 */
 	public MaskedTextFieldBuilder maskedTextField(Attribute<String> attribute) {
-		ValueAttributeDefinition<String> attributeDefinition = definition(attribute);
+		AttributeDefinition<String> attributeDefinition = definition(attribute);
 
 		return Components.maskedTextField()
 						.toolTipText(attributeDefinition.description().orElse(null));
@@ -617,7 +619,7 @@ public final class EntityComponents {
 	 * @return a {@link FileInputPanel.Builder}
 	 */
 	public ComponentValueBuilder<FileInputPanel, byte[], FileInputPanel.Builder<byte[]>> byteArrayInputPanel(Attribute<byte[]> attribute) {
-		ValueAttributeDefinition<byte[]> attributeDefinition = definition(attribute);
+		AttributeDefinition<byte[]> attributeDefinition = definition(attribute);
 
 		return Components.byteArrayInputPanel()
 						.toolTipText(attributeDefinition.description().orElse(null));
@@ -631,13 +633,38 @@ public final class EntityComponents {
 		return new EntityComponents(entityDefinition);
 	}
 
-	private <T> ValueAttributeDefinition<T> definition(Attribute<T> attribute) {
-		AttributeDefinition<T> definition = entityDefinition.attributes().definition(attribute);
-		if (!(definition instanceof ValueAttributeDefinition<T>)) {
-			throw new IllegalArgumentException("Value based attribute expected: " + attribute);
+	private <T> AttributeDefinition<T> definition(Attribute<T> attribute) {
+		return entityDefinition.attributes().definition(attribute);
+	}
+
+	private static boolean nullable(AttributeDefinition<?> attributeDefinition) {
+		return attributeDefinition instanceof ValueAttributeDefinition<?> &&
+						((ValueAttributeDefinition<?>) attributeDefinition).nullable();
+	}
+
+	private static boolean itemBased(AttributeDefinition<?> definition) {
+		return definition instanceof ValueAttributeDefinition<?> && !((ValueAttributeDefinition<?>) definition).items().isEmpty();
+	}
+
+	private static @Nullable Number minimumValue(AttributeDefinition<?> attributeDefinition) {
+		if (!(attributeDefinition instanceof ValueAttributeDefinition<?>)) {
+			return null;
 		}
 
-		return (ValueAttributeDefinition<T>) definition;
+		return ((ValueAttributeDefinition<?>) attributeDefinition).minimum().orElse(null);
+	}
+
+	private static @Nullable Number maximumValue(AttributeDefinition<?> attributeDefinition) {
+		if (!(attributeDefinition instanceof ValueAttributeDefinition<?>)) {
+			return null;
+		}
+
+		return ((ValueAttributeDefinition<?>) attributeDefinition).maximum().orElse(null);
+	}
+
+	private static int maximumLength(AttributeDefinition<?> attributeDefinition) {
+		return attributeDefinition instanceof ValueAttributeDefinition<?> ?
+						((ValueAttributeDefinition<?>) attributeDefinition).maximumLength() : -1;
 	}
 
 	private static <T> FilterComboBoxModel<T> createEnumComboBoxModel(Attribute<T> attribute, boolean nullable) {